ZigZag: read pivot values from the point, switch pivots to double
OptimizeList recomputed Math.Max(Open, Close) from the candle on every
candle, while the point already holds exactly that value: a point's Value
is the low of its candle for an 'L' and the high for an 'H'. The L/H/L and
H/L/H alternation check above it made the if/else redundant.

Pivot values are now double instead of decimal - a pivot only ever answers
"is this point higher or lower", it is not money that has to add up. The
conversion preserves both ordering and equality for prices of this
magnitude. Values are cast back to decimal where a pivot leaves the trend
world as a price (CryptoTrendData, StructureEvent, DlzAdmin, chart levels);
four casts in the chart code became superfluous.

Throughput: 6.70 -> 2.34 us/candle, roughly 210 s off a 48.6M candle
replay. Pinned zigzag outcome unchanged apart from trailing zeros in the
formatting; all 375 tests pass.

+++ CryptoScanner.Chart/ViewModels/Chart/FibRetracement.cs
@@ -62,7 +62,7 @@ public class FibRetracement
                 last = indicator.LastSwingLow;
                 //fibRetracement = Retracement2(indicator.LastSwingHigh.Value, indicator.LastSwingLow.Value);
             }
-            fibRetracement = RetracementX(side, indicator.LastSwingLow.Value, indicator.LastSwingHigh.Value);
+            fibRetracement = RetracementX(side, (decimal)indicator.LastSwingLow.Value, (decimal)indicator.LastSwingHigh.Value);
 
             CryptoSymbolInterval symbolInterval = symbol.GetSymbolInterval(interval.IntervalPeriod);
             CandleTime start = last.Candle.OpenTime + interval.Duration;
